Forum manager and sync manager ajax and nonajax improvements.
