+++ Resources/public/js/Zikula.Dizkus.Sync.js
@@ -1,17 +1,14 @@
 /**
- * Zikula.Dizkus.Sync.js
+ * Zikula.Dizkus.ForumTree.js
  *
  * $ based JS
  */
-
-// in case not exists
 var Zikula = Zikula || {};
 // Dizkus namespace
 Zikula.Dizkus = Zikula.Dizkus || {};
 (function ($, Routing, Translator) {
-//Upgrade module
     Zikula.Dizkus.Sync = (function () {
-// Init
+        // Init
         var data = {
             log: null
         };
@@ -25,15 +22,17 @@ Zikula.Dizkus = Zikula.Dizkus || {};
         ;
         function init()
         {
-            log(Translator.__('Sync init start.'));
-            initTree();
-            log(Translator.__('Sync init done.'));
+            log(Translator.__('Forum tree init start.'));
+
+            startListeners();
+
+            log(Translator.__('Forum tree init done.'));
         }
         ;
         function readSettings()
         {
 //            settings.ajax_timeout = parseInt($("#ajax_timeout").val());
-            log(Translator.__('Sync settings updated.'));
+            log(Translator.__('Forum tree settings updated.'));
         }
         ;
         function log(log)
@@ -52,107 +51,63 @@ Zikula.Dizkus = Zikula.Dizkus || {};
             }
         }
         ;
-        // initialise jstree's
-        /*
-         * forum Tree
-         */
-        function initTree() {
-            $('#sync_tree')
-                    .jstree({})
-                    .on('after_close.jstree', function (e, data) {
-                    })
-                    .on('after_open.jstree', function (e, data) {
-                    })
-                    .on('ready.jstree', function (e, data) {
-                    })
-                    .on('changed.jstree', function (e, data) {
-                    })
-                    .on('select_node.jstree', function (e, data) {
-                        e.preventDefault();
-                        e.stopImmediatePropagation();
-                        return false;
-                    });
+        function startListeners()
+        {
+            $(".js-switch").bind('click', function (e) {
+                e.preventDefault();
+                e.stopPropagation();
+                log('work');
+                prepareTree().done(initTree());
+                $(this).addClass('disabled');
+            });
 
-
-
-            $('#sync_tree')
-                    .jstree(true)
-                    .settings
-                    .core
-                    .data = {
-                        'url': Routing.generate('zikuladizkusmodule_sync_forumtree'),
-                        "dataType": "json",
-                        "dataFilter": function (response) {
-                            res = JSON.parse(response);
-                            data.tree = res.tree;
-                            data.total = res.total;
-                            console.log(data);
-//                            data.excluded = res.excluded;
-//                            data.total.topics = data.total.topics - data.excluded.topics.length;
-//                            data.total.posts = data.total.posts - data.excluded.posts.length;
-//                            $("#forum_legend").prepend($(".progress").first().clone());
-//                            $("#forum_legend").find('.progress').removeClass('hide');
-//                            $("#forum_legend")
-//                                    .find('.info')
-//                                    .text(Translator.__('Forums (+categories):') + ' '
-//                                            + data.total.forums + ' '
-//                                            + Translator.__('Topics:') + ' '
-//                                            + data.total.topics + ' '
-//                                            + Translator.__('Posts:') + ' '
-//                                            + data.total.posts)
-//                                    .css('color', '#000');
-//                            log(Translator.__('To import:') + $("#forum_legend").find('.info').text());
-//                            log(Translator.__('Forum tree loaded.'));
-//                            if (data.total.current > 1) {
-//                                log(Translator.__('Error! Content detected in target tables.'));
-//                                $("#remove_forum_tree").removeClass('hide disabled');
-//                            } else {
-//                                $("#recover_forum_tree").removeClass('btn-primary').addClass('btn-success');
-//                                $("#import_forum_tree").removeClass('disabled').addClass('btn-primary');
-//                            }
-//                            $('<p id="rejected_topics_items" title="' + Translator.__('Rejected topics') + '" class="text-muted small"></p>')
-//                                    .prepend('<i class="fa fa-hashtag " aria-hidden="true"></i> ' + Translator.__('Rejected topics') + ' <br />')
-//                                    .appendTo($('#import_rejected'));
-//                            $.each(data.excluded.topics, function (index, item) {
-//                                var reason = decodeRejectedReason(item.reason);
-//                                $('#rejected_topics_items')
-//                                        .append('<span title="' + reason + '" class="text-muted small">\n\
-//                            <i class="fa fa-hashtag text-danger" aria-hidden="true"></i><span class="rejected_id">' + item.id + ' ' + reason + ' </span></span>');
-//                            });
-//                            $('<p id="rejected_posts_items" title="' + Translator.__('Rejected posts') + '" class="text-muted small"></p>')
-//                                    .prepend('<i class="fa fa-hashtag " aria-hidden="true"></i> ' + Translator.__('Rejected posts') + ' <br />')
-//                                    .appendTo($('#import_rejected'));
-//                            $.each(data.excluded.posts, function (index, item) {
-//                                var reason = decodeRejectedReason(item.reason);
-//                                $('#rejected_posts_items')
-//                                        .append('<span title="' + reason + '" class="text-muted small">\n\
-//                            <i class="fa fa-hashtag text-danger" aria-hidden="true"></i><span class="rejected_id">' + item.id + ' ' + reason + ' </span></span>');
-//                            });
-
-                            return JSON.stringify(prepateTreeData(data.tree));
-                            ;
-                        }
-                        ,
-                        'data': function (node) {
-                            return {'id': node.id, 'text': node.name};
-                        }
-                    };
-//            $(".jstree-node").bind('click', function (e) {
-//                e.preventDefault();
-//                e.stopPropagation();
-//                return false;
-//            });
-//            $(".jstree-icon").bind('click', function (e) {
-//                e.preventDefault();
-//                e.stopPropagation();
-//                return false;
-//            });
+            log(Translator.__('Forum tree listeners started.'));
         }
         ;
-//        function getForumTree() {
-//            $("#forum_tree").jstree("open_node", $("#forum_tree_root"));
-//        }
-//
+        function prepareTree()
+        {
+            log(Translator.__('Tree preparation start.'));
+            var def = $.Deferred();
+            $('.noajax-actions').remove();
+            // define the promise that is called once all elements are removed
+            $('.noajax-actions').promise().done(function () {
+                def.resolve();
+            });
+            $('.forum-title').attr('class', 'forum-title');
+            log(Translator.__('Tree preparation done.'));
+            return def.promise();
+        }
+        ;
+        /*
+         * Forum Tree
+         *
+         * initialise jstree's
+         */
+        function initTree() {
+            $('#forum_tree')
+                    .jstree({
+                        'core': {
+                            'check_callback': true
+                        }
+                        ,
+                        "plugins": [
+                            "checkbox",
+                            "dnd",
+                            "contextmenu"
+                        ],
+                        'checkbox': {
+                            'keep_selected_style': false,
+                            'three_state': false,
+                            'cascade': ''
+                        },
+                        'contextmenu': {
+                            select_node: false,
+                            items: itemMenu
+                        }
+                    });
+        }
+        ;
+
         function prepateTreeData(tree) {
             tree = prepareNodeData(tree);
             return tree;
@@ -166,9 +121,46 @@ Zikula.Dizkus = Zikula.Dizkus || {};
             return node;
         }
 
+        function itemMenu(node)
+        {
+            log(node);
+            var items = {
+                "Modify": {
+                    "label": Translator.__('Modify'),
+                    "icon": 'fa fa-pencil',
+                    "action": function (obj) {
+                        log(obj);
+                    }
+                },
+                "Lock": {
+                    "label": Translator.__('Lock'),
+                    "icon": 'fa fa-lock',
+                    "action": function (obj) {
+                        log(obj);
+                    }
+                },
+                "Sync": {
+                    "label": Translator.__('Sync'),
+                    "icon": 'fa fa-refresh',
+                    "action": function (obj) {
+                        log(obj);
+                    }
+                },
+                "Delete": {
+                    "label": Translator.__('Delete'),
+                    "icon": 'fa fa-trash',
+                    "separator_before": true,
+                    "action": function (obj) {
+                        log(obj);
+                    }
+                }
+            };
+
+            return items;
+        }
+
         //ajax util
         function importAjax(url, data) {
-//            console.log(data);
             return $.ajax({
                 type: 'POST',
                 url: url,
@@ -179,11 +171,12 @@ Zikula.Dizkus = Zikula.Dizkus || {};
             });
         }
 
-        //return this and init when ready
+        //expose actions
         return {
             init: init
         };
     })();
+    //autoinit
     $(function () {
         Zikula.Dizkus.Sync.init();
     });
