--- Resources/views/Admin/tree.html.twig
@@ -1,109 +1,0 @@
-{{ adminHeader() }}
-<link href="{{ pageAddAsset('stylesheet', zasset('jstree/jstree-built.css'), 98)  }}" rel="stylesheet" type="text/css" />
-{# @todo move to separate file #}
-{% macro drawTree(items) %}
-    {% for item in items %}
-        <li class="clearfix">
-            <div class="noajax-actions col-md-1">
-                {% if not loop.first %}
-                    <a class="btn btn-xs tooltips" href="{{ path('zikuladizkusmodule_admin_changeforumorder', {'forum': item.forum_id, 'action': 'moveUp'})}}" title='{{ __("move up")}}'>
-                        <i class="fa fa-arrow-up"></i>
-                    </a>
-                {% endif %}
-                {% if not loop.last %}
-                    <a class="btn btn-xs tooltips" href="{{ path('zikuladizkusmodule_admin_changeforumorder', {'forum': item.forum_id, 'action': 'moveDown'})}}"  title='{{ __("move down")}}'>
-                        <i class="fa fa-arrow-down"></i>
-                    </a>
-                {% endif %}
-            </div>
-            <span class="tooltips" title='{{ __("Forum #{item.name}")}}'>
-                {% if item.name is constant('Zikula\\DizkusModule\\Entity\\ForumEntity::ROOTNAME') %}
-                    <strong>{{ __('Forum index') }}</strong>
-                {% else %}
-                    {{ item.name }}
-                {% endif %}
-            </span>
-            <span class="noajax-actions text-right">
-                <a class="btn btn-xs tooltips" href="{{ path('zikuladizkusmodule_admin_modifyforum', {'id': item.forum_id})}}" title='{{ __("edit forum #{item.name}")}}'><i class="fa fa-pencil"></i></a>
-                <a class="btn btn-xs tooltips" href="{{ path('zikuladizkusmodule_forum_viewforum', {'forum': item.forum_id})}}" title='{{ __("view forum #{item.name}")}}'><i class="fa fa-eye fa-blue"></i></a>
-                <a class="btn btn-xs tooltips" href="{{ path('zikuladizkusmodule_admin_deleteforum', {'id': item.forum_id})}}" title='{{ __("delete forum #{item.name}")}}'><i class="fa fa-trash-o fa-red"></i></a>
-            </span>
-            {% if item.__children > 0 %}
-                <ul class="" style="list-style:none;">
-                    {{ _self.drawTree(item.__children)}}
-                </ul>
-            {% endif %}
-        </li>
-    {% endfor %}
-{% endmacro %}
-
-<div id="dizkus_admin">
-    {% if importHelper.isUpgrading() %}
-        <div id="upgrade_notice" class="alert alert-warning">
-            <div class="row">
-                <div class="col-md-12">
-                    <i class="close" data-dismiss="alert">×</i>
-                    <strong><i class="fa fa-info-circle"></i> {{ __('Module is in upgrade state.')}}&nbsp;{{ __('Version')}}&nbsp;{{ importHelper.isUpgrading() | replace('_', ' ')}} </strong>
-                    <a class="btn btn-success" href="{{ path('zikuladizkusmodule_import_import')}}" title="{{ __('Finish data import.')}}">
-                        <span class="fa fa-cloud-download"></span>&nbsp;{{ __('Finish data import.')}}
-                    </a>
-                </div>
-            </div>
-        </div>
-    {% else %}
-    {% endif %}
-    <div class="panel panel-default">
-        <div class="panel-heading">
-            <h3><span class="fa fa-sitemap"></span> {{ __('Forum tree')}}</h3>
-        </div>
-        <div class="list-group-item list-group-item-warning" style="">
-            <a class="btn btn-primary" href="{{ path('zikuladizkusmodule_admin_modifyforum')}}" title="{{ __('Create a new forum')}}">
-                <span class="fa fa-comments"></span>&nbsp;{{ __('Create a new forum')}}
-            </a>
-            <a class="btn btn-default pull-right" href="{{ path('zikuladizkusmodule_sync_sync')}}" title="{{ __('Recalculate post and topics totals')}}">
-                <span class="fa fa-cogs"></span>&nbsp;{{ __('Recalculate post and topics totals')}}
-            </a>
-        </div>
-        {% if status is not same as(true) %}
-            <div class="list-group-item list-group-item-danger ">
-                <div class="row">
-                    <div class="col-md-12">
-                        <h4> {{ __('Tree errors found:') }}</h4>
-                        {% for err in status %}
-                            <span class="muted text-danger small"><i class="fa fa-exclamation-triangle" ></i> &nbsp;{{ err }}</span>
-                        {% endfor %}
-                    </div>
-                </div>
-            </div>
-        {% endif %}
-        <div class="list-group-item list-group-item-info">
-            <div class="row">
-                <div class="col-md-11">    
-                    {{ __('Tree')}}
-                </div>
-                <div class="col-md-1">
-                    {{ __('Actions')}}   
-                </div>
-            </div>
-        </div>
-        <div class="list-group-item">
-            <div class="row">
-                <div id="forum_tree" class="col-md-12">
-                    <ul class="list-unstyled">
-                        {{ _self.drawTree(tree)}}
-                    </ul>
-                </div>
-            </div>
-        </div>
-        <div class="list-group-item list-group-item-info">
-            <div class="row">
-                <div class="col-md-12">
-
-                </div>
-            </div>
-        </div>
-    </div>
-    {#    {{ dump() }}#}
-    {{ pageAddAsset('javascript', zasset('jstree/jstree-built.js'), 98)}}
-    {{ pageAddAsset('javascript', zasset('@ZikulaDizkusModule:js/Zikula.Dizkus.ForumTree.js'))}}
-    {{ adminFooter()}}
